EditMode enum refactor
