+++ src/components/common/componentEnums.ts
@@ -1,0 +1,4 @@
+export enum EditMode {
+    View,
+    Edit
+}
