+++ src/components/molecules/buttons/index.ts
@@ -2,7 +2,7 @@ export { AddButton } from "./AddButton";
 export { ArchiveButton } from "./ArchiveButton";
 export { CancelButton } from "./CancelButton";
 export { DoneButton } from "./DoneButton";
-export { EditButton, EditMode } from "./EditButton";
+export { EditButton } from "./EditButton";
 export { FrameCloseButton } from "./FrameCloseButton";
 export { HomeButton } from "./HomeButton";
 export { RefreshButton } from "./RefreshButton";
