+++ src/index.ts
@@ -4,6 +4,9 @@ export { FEATURE_TOGGLE_LIST } from "./constants/defaultFeatureToggles";
 export { PushState } from "./reducers/types";
 export { rootReducerInitialState } from "./reducers/rootReducer";
 
+// components - common
+export * from "./components/common/index";
+
 // components - inputs - atoms
 export { Checkbox } from "./components/atoms/inputs/Checkbox";
 export { StandardInput } from "./components/atoms/inputs/StandardInput";
