+++ src/views/ReviewView.tsx
@@ -5,7 +5,7 @@ import * as React from "react";
 import { TopMenuPanelContainer } from "../containers/TopMenuPanelContainer";
 
 // consts/enums
-import { EditMode } from "../components/molecules/buttons/EditButton";
+import { EditMode } from "../components/common/componentEnums";
 
 export interface ReviewViewStateProps {
     editMode: EditMode;
